+++ job-factory.go
@@ -32,65 +32,3 @@ func JobFactory(jobRes json.RawMessage) (*Job, error) {
 
 	return &job, nil
 }
-
-// -------- DockerHostTaskJob --------
-
-type dockerHostTaskJob struct {
-	Command string `json:"command"`
-}
-
-type DockerHostTaskJob struct {
-	*BasicJob
-	*dockerHostTaskJob
-}
-
-func (job *DockerHostTaskJob) UnmarshalJSON(b []byte) error {
-	var bj BasicJob
-	if err := json.Unmarshal(b, &bj); err == nil {
-		var j dockerHostTaskJob
-		if err := json.Unmarshal(bj.basicJob.ParamsRaw, &j); err != nil {
-			return err
-		}
-		*job = DockerHostTaskJob{BasicJob: &bj, dockerHostTaskJob: &j}
-		return nil
-	} else {
-		return err
-	}
-}
-
-func (job DockerHostTaskJob) GetBasicJob() BasicJob {
-	return *job.BasicJob
-}
-
-// -------- DockerServiceTaskJob --------
-
-type dockerServiceTaskJob struct {
-	Task        string `json:"task"`
-	ServiceName string `json:"service_name"`
-	PrivateIp   string `json:"private_ip"`
-}
-
-type DockerServiceTaskJob struct {
-	*BasicJob
-	*dockerServiceTaskJob
-}
-
-func (job *DockerServiceTaskJob) UnmarshalJSON(b []byte) error {
-	var bj BasicJob
-	if err := json.Unmarshal(b, &bj); err == nil {
-		var j dockerServiceTaskJob
-		if err := json.Unmarshal(bj.basicJob.ParamsRaw, &j); err != nil {
-			return err
-		}
-		*job = DockerServiceTaskJob{BasicJob: &bj, dockerServiceTaskJob: &j}
-		return nil
-	} else {
-		return err
-	}
-}
-
-func (job DockerServiceTaskJob) GetBasicJob() BasicJob {
-	return *job.BasicJob
-}
-
-// ----------------
